+++ fw18_0279-day4/Search.html
@@ -106,7 +106,7 @@
        }
      #K{
         height:120px ;
-       
+        cursor: pointer;
         border: 1px solid black;
         margin-left: 10px;
      }
@@ -124,7 +124,7 @@
 
     <h3 style="margin-left:12% ;font-family: arial;">Browse Common Foods</h3>
     <hr>
-    <div id="Search"></div>
+    <div id="Search" ></div>
    
 </body>
 </html>
